# Resume-cursor/openapi_live.json
{"openapi":"3.1.0","info":{"title":"简历审查 Agent","version":"0.1.0"},"paths":{"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/tenants":{"get":{"tags":["admin"],"summary":"List Tenants","operationId":"list_tenants_api_admin_tenants_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TenantOut"},"type":"array","title":"Response List Tenants Api Admin Tenants Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["admin"],"summary":"Create Tenant","operationId":"create_tenant_api_admin_tenants_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/admin/tenants/{tenant_id}":{"patch":{"tags":["admin"],"summary":"Patch Tenant","operationId":"patch_tenant_api_admin_tenants__tenant_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"tenant_id","in":"path","required":true,"schema":{"type":"string","title":"Tenant Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantPatch"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/invites":{"get":{"tags":["invites"],"summary":"List Invites","operationId":"list_invites_api_invites_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InviteOut"},"type":"array","title":"Response List Invites Api Invites Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["invites"],"summary":"Create Tenant Invite","operationId":"create_tenant_invite_api_invites_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/reviews/extract":{"post":{"tags":["reviews"],"summary":"Preview Extract","operationId":"preview_extract_api_reviews_extract_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_preview_extract_api_reviews_extract_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtractPreviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/reviews":{"get":{"tags":["reviews"],"summary":"List Reviews","operationId":"list_reviews_api_reviews_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ReviewListItem"},"type":"array","title":"Response List Reviews Api Reviews Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["reviews"],"summary":"Create Review","operationId":"create_review_api_reviews_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_review_api_reviews_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/reviews/{review_id}":{"get":{"tags":["reviews"],"summary":"Get Review","operationId":"get_review_api_reviews__review_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"string","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/report":{"get":{"summary":"Report Page","operationId":"report_page_report_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_create_review_api_reviews_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"job_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Title"},"job_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Description"}},"type":"object","required":["file"],"title":"Body_create_review_api_reviews_post"},"Body_preview_extract_api_reviews_extract_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_preview_extract_api_reviews_extract_post"},"ExtractPreviewResponse":{"properties":{"filename":{"type":"string","title":"Filename"},"anchored_text":{"type":"string","title":"Anchored Text"},"char_count":{"type":"integer","title":"Char Count"},"sentence_count":{"type":"integer","title":"Sentence Count"},"section_count":{"type":"integer","title":"Section Count"},"fallback_sentence_count":{"type":"integer","title":"Fallback Sentence Count","default":0},"fallback_ratio":{"type":"number","title":"Fallback Ratio","default":0.0},"layout":{"type":"string","title":"Layout","default":"single"},"page_count":{"type":"integer","title":"Page Count","default":0}},"type":"object","required":["filename","anchored_text","char_count","sentence_count","section_count"],"title":"ExtractPreviewResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InviteCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"role":{"$ref":"#/components/schemas/UserRole","default":"member"}},"type":"object","required":["email"],"title":"InviteCreate"},"InviteOut":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"role":{"$ref":"#/components/schemas/UserRole"},"token":{"type":"string","title":"Token"},"expires_at":{"type":"string","format":"date-time","title":"Expires At"},"used_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Used At"}},"type":"object","required":["id","email","role","token","expires_at","used_at"],"title":"InviteOut"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":6,"title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"RegisterRequest":{"properties":{"token":{"type":"string","maxLength":64,"minLength":8,"title":"Token"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":6,"title":"Password"}},"type":"object","required":["token","email","password"],"title":"RegisterRequest"},"ReviewCreateResponse":{"properties":{"id":{"type":"string","title":"Id"},"agent_key":{"type":"string","title":"Agent Key"},"status":{"type":"string","title":"Status"},"overall_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Overall Score"},"weights":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Weights"},"report":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Report"},"filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filename"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["id","agent_key","status"],"title":"ReviewCreateResponse"},"ReviewListItem":{"properties":{"id":{"type":"string","title":"Id"},"status":{"type":"string","title":"Status"},"overall_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Overall Score"},"filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filename"},"job_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Title"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","status","overall_score","filename","job_title","created_at"],"title":"ReviewListItem"},"TenantCreate":{"properties":{"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"slug":{"type":"string","maxLength":80,"minLength":2,"pattern":"^[a-z0-9_-]+$","title":"Slug"},"admin_email":{"type":"string","format":"email","title":"Admin Email"}},"type":"object","required":["name","slug","admin_email"],"title":"TenantCreate"},"TenantOut":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"slug":{"type":"string","title":"Slug"},"is_active":{"type":"boolean","title":"Is Active"},"invite_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Invite Token"},"invite_email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Invite Email"}},"type":"object","required":["id","name","slug","is_active"],"title":"TenantOut"},"TenantPatch":{"properties":{"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["is_active"],"title":"TenantPatch"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserOut":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"role":{"$ref":"#/components/schemas/UserRole"},"tenant_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tenant Id"},"tenant_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tenant Name"}},"type":"object","required":["id","email","role","tenant_id"],"title":"UserOut"},"UserRole":{"type":"string","enum":["platform_admin","tenant_admin","member"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}